connecting actions to page controller
